add graph_node_event_listener

+++ bridge/validator/src/substrate_transactions.rs
@@ -16,7 +16,7 @@ pub fn mint(
     message_id: primitives::H256,
     from: primitives::H160,
     to: AccountId,
-    amount: u64,
+    amount: u128,
 ) {
     let xthex = build_mint(
         &sub_api,
@@ -68,7 +68,7 @@ pub fn build_mint(
     message_id: H256,
     from: H160,
     to: AccountId,
-    amount: u64,
+    amount: u128,
 ) -> String {
     let signer_index = signer_index(sub_api, &signer);
     let genesis_hash = sub_api.genesis_hash.expect("can not get genesiss hash");
